CLDR-16786 update as per review comments. Also moved test into a file that is run in TestAll!
